changed split (where used as a noun) to feature flag, updated API doc links
Updated postman collection file to reflect new naming in the Split product and docs that was announced May 15, 2023 in https://www.split.io/blog/new-split-ui-made-by-developers-for-developers/  

tl;dr: "splits" are now referred to as "feature flags" 

Updated all mentions of "split" (where used as a noun in titles, narrative text, and placeholder names) to "feature flag."

Did not change "split" if it appeared in an API path name or as a variable (i.e., as the "name" in a name:value pair).

Updated links to API docs where the doc now has "feature-flag" in the path, rather than "split."  
Example: https://docs.split.io/reference/split-definition 
was fixed to be https://docs.split.io/reference/feature-flag-definition

Repaired links to API docs in the description field, where there was previously an anchor (#) instead of a slash (/) in the path to the doc page.  Example: https://docs.split.io/reference#get-workspaces 
was fixed to be https://docs.split.io/reference/get-workspaces.