Update Harness endpoint response examples: Re-run all Harness endpoints and save response examples
Changes include:
* Add response headers
* Adjust order of collection variables
* Add references to collection variables in endpoints' request body and params

Note: We recently placed this collection inside a team account. That changed the URL to the collection on the Postman network. Here is the updated URL: https://www.postman.com/harness-fme-enablement/harness-fme/collection/hyphfpd/before-and-after-apis-for-split-admins

Related documentation on Harness Developer Hub is here: https://developer.harness.io/docs/feature-management-experimentation/split-to-harness/api-for-split-admins